Restructure lib files
