+++ src/lib/docker/generateTag.spec.ts
